+++ classes/Utils/Cron.php
@@ -19,7 +19,7 @@ class PUM_Utils_Cron {
 	 */
 	public function __construct() {
 		add_filter( 'cron_schedules', array( $this, 'add_schedules' ) );
-		add_action( 'wp', array( $this, 'schedule_Events' ) );
+		add_action( 'wp', array( $this, 'schedule_events' ) );
 	}
 
 	/**