+++ CHANGELOG.md
@@ -1,6 +1,20 @@
+### v1.11.0 - 06/25/2020
+* Feature: Add new floating bar theme.
+* Feature: New guided tour of popup editor for first time users.
+* Feature: New href attribute on the popup_close shortcode for when setting the shortcode to use the `a` tag.
+* Fix: Shortcode popup_close tag attribute not functioning properly.
+* Tweak: Change popup_close tag setting to use a drop-down for more easily selecting which tag to use.
+* Tweak: Improve explanation of cookies in new trigger modal.
+* Tweak: Ensure all admin pages have visible page headings.
+* Tweak: Simplify name and title fields in popup editor.
+* Tweak: Add popup ID to quick links on All Popups.
+* Tweak: Move CSS and JS for our admin bar node to external file.
+* Tweak: Add our new optional telemetry system.
+
 ### v1.10.2 - 06/09/2020
 * Fix: Popup Settings not working when WP Forms is active without forms
 * Fix: Missing closing div in new [popup_cookie] shortcode.
+* Fix: Shortcode popup_close tag attribute not functioning properly.
 
 ### v1.10.1 - 04/21/2020
 * Fix: Typo in filter name caused extra p tags.